requirements.txt
pillow
python-dotenv
requests
